Moved AppContext.js into a new Context directory. Added Admin and Coach directories to hold components used for specific dashboards
