+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
 import './App.css';
-import { AppContext } from './components/AppContext';
+import { AppContext } from './components/Context/AppContext';
 import LandingPage from './components/LandingPage';
 
 import AdminDashboard from './components/Dashboards/AdminDashboard';
