+++ frontend/src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css';
 
 import App from './App';
-import AppProvider from './components/AppContext';
+import AppProvider from './components/Context/AppContext';
 
 ReactDOM.render(
   <AppProvider>
